refactor: Rename networking package to observer
