+++ arcade-nametags/src/main/kotlin/net/casual/arcade/nametags/Nametag.kt
@@ -5,7 +5,7 @@
 package net.casual.arcade.nametags
 
 import net.casual.arcade.nametags.virtual.NametagHeight
-import net.casual.arcade.networking.observer.Observer
+import net.casual.arcade.observer.Observer
 import net.casual.arcade.utils.TimeUtils.Ticks
 import net.casual.arcade.utils.color.ColorARGB
 import net.casual.arcade.utils.time.MinecraftTimeDuration
