+++ arcade-observers/src/main/kotlin/net/casual/arcade/observer/ArcadeObservers.kt
@@ -2,9 +2,9 @@
  * Copyright (c) 2026 senseiwells
  * Licensed under the MIT License. See LICENSE file in the project root for details.
  */
-package net.casual.arcade.networking
+package net.casual.arcade.observer
 
-import net.casual.arcade.networking.extensions.PlayerObserverExtension
+import net.casual.arcade.observer.extensions.PlayerObserverExtension
 import net.fabricmc.api.ModInitializer
 
 public object ArcadeObservers: ModInitializer {
