+++ arcade-observers/src/main/kotlin/net/casual/arcade/observer/Observer.kt
@@ -2,9 +2,9 @@
  * Copyright (c) 2026 senseiwells
  * Licensed under the MIT License. See LICENSE file in the project root for details.
  */
-package net.casual.arcade.networking.observer
+package net.casual.arcade.observer
 
-import net.casual.arcade.networking.packet.PacketSender
+import net.casual.arcade.observer.packet.PacketSender
 import net.casual.arcade.utils.math.location.LocationWithLevel
 import net.minecraft.network.protocol.Packet
 import net.minecraft.server.level.ServerLevel
