+++ arcade-observers/src/main/kotlin/net/casual/arcade/observer/PlayerObserver.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2026 senseiwells
  * Licensed under the MIT License. See LICENSE file in the project root for details.
  */
-package net.casual.arcade.networking.observer
+package net.casual.arcade.observer
 
 import net.casual.arcade.utils.math.location.LocationWithLevel
 import net.casual.arcade.utils.math.location.locationWithLevel
