+++ arcade-observers/src/main/kotlin/net/casual/arcade/observer/packet/PacketSender.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2026 senseiwells
  * Licensed under the MIT License. See LICENSE file in the project root for details.
  */
-package net.casual.arcade.networking.packet
+package net.casual.arcade.observer.packet
 
 import net.minecraft.network.protocol.Packet
