+++ arcade-observers/src/main/kotlin/net/casual/arcade/observer/tracker/SimpleObserverTracker.kt
@@ -2,10 +2,10 @@
  * Copyright (c) 2026 senseiwells
  * Licensed under the MIT License. See LICENSE file in the project root for details.
  */
-package net.casual.arcade.networking.observer.tracker
+package net.casual.arcade.observer.tracker
 
 import it.unimi.dsi.fastutil.objects.ObjectLinkedOpenHashSet
-import net.casual.arcade.networking.observer.Observer
+import net.casual.arcade.observer.Observer
 
 public class SimpleObserverTracker: ObserverTracker {
     private val tracking = ObjectLinkedOpenHashSet<Observer>()
