+++ arcade-observers/src/main/kotlin/net/casual/arcade/observer/utils/ObserverUtils.kt
@@ -2,11 +2,11 @@
  * Copyright (c) 2026 senseiwells
  * Licensed under the MIT License. See LICENSE file in the project root for details.
  */
-package net.casual.arcade.networking.utils
+package net.casual.arcade.observer.utils
 
-import net.casual.arcade.networking.observer.Observer
-import net.casual.arcade.networking.observer.PlayerObserver
-import net.casual.arcade.networking.extensions.PlayerObserverExtension.Companion.observerExtension
+import net.casual.arcade.observer.Observer
+import net.casual.arcade.observer.PlayerObserver
+import net.casual.arcade.observer.extensions.PlayerObserverExtension.Companion.observerExtension
 import net.minecraft.server.level.ServerPlayer
 
 public fun Observer.asPlayerOrNull(): ServerPlayer? {
