+++ arcade-observers/src/main/resources/arcade-observers.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "net.casual.arcade.networking.mixins",
+  "package": "net.casual.arcade.observer.mixins",
   "compatibilityLevel": "JAVA_17",
   "mixinextras": {
     "minVersion": "0.5.0"
