+++ arcade-virtual-entities/src/main/kotlin/net/casual/arcade/virtual/entity/SimpleParentVirtualEntity.kt
@@ -5,8 +5,8 @@
 package net.casual.arcade.virtual.entity
 
 import it.unimi.dsi.fastutil.objects.ObjectLinkedOpenHashSet
-import net.casual.arcade.networking.observer.tracker.ObserverTracker
-import net.casual.arcade.networking.observer.tracker.SimpleObserverTracker
+import net.casual.arcade.observer.tracker.ObserverTracker
+import net.casual.arcade.observer.tracker.SimpleObserverTracker
 import net.casual.arcade.virtual.entity.attachment.VirtualEntityAttachment
 import net.casual.arcade.virtual.entity.attachment.anchor.AttachmentAnchor
 import net.casual.arcade.virtual.entity.location.VirtualPosition
