+++ arcade-virtual-entities/src/main/kotlin/net/casual/arcade/virtual/entity/attachment/RootVirtualEntityAttachment.kt
@@ -4,9 +4,9 @@
  */
 package net.casual.arcade.virtual.entity.attachment
 
-import net.casual.arcade.networking.observer.Observer
-import net.casual.arcade.networking.observer.PlayerObserver
-import net.casual.arcade.networking.packet.PacketSender
+import net.casual.arcade.observer.Observer
+import net.casual.arcade.observer.PlayerObserver
+import net.casual.arcade.observer.packet.PacketSender
 import net.casual.arcade.virtual.entity.extensions.PlayerAttachmentObserverExtension.Companion.attachmentObserverExtension
 import net.casual.arcade.virtual.entity.utils.VirtualEntityPacketCollector
 import net.casual.arcade.virtual.entity.utils.sendBundledSpawnPackets
