+++ arcade-virtual-entities/src/main/kotlin/net/casual/arcade/virtual/entity/collision/CollisionCubeVirtualEntity.kt
@@ -4,9 +4,9 @@
  */
 package net.casual.arcade.virtual.entity.collision
 
-import net.casual.arcade.networking.observer.Observer
-import net.casual.arcade.networking.observer.tracker.ObserverTracker
-import net.casual.arcade.networking.packet.PacketSender
+import net.casual.arcade.observer.Observer
+import net.casual.arcade.observer.tracker.ObserverTracker
+import net.casual.arcade.observer.packet.PacketSender
 import net.casual.arcade.utils.ClientboundSetPassengersPacket
 import net.casual.arcade.virtual.entity.SimpleParentVirtualEntity
 import net.casual.arcade.virtual.entity.SimpleVirtualEntity
