+++ arcade-virtual-entities/src/main/kotlin/net/casual/arcade/virtual/entity/data/PlayerSpecificEntityData.kt
@@ -8,8 +8,6 @@ import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap
 import it.unimi.dsi.fastutil.ints.IntOpenHashSet
 import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap
 import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet
-import net.casual.arcade.networking.observer.Observer
-import net.casual.arcade.networking.utils.asPlayerOrNull
 import net.minecraft.network.syncher.EntityDataAccessor
 import net.minecraft.network.syncher.SynchedEntityData
 import net.minecraft.world.entity.EntityType
