+++ arcade-virtual-entities/src/main/kotlin/net/casual/arcade/virtual/entity/display/SimpleVirtualDisplay.kt
@@ -5,8 +5,8 @@
 package net.casual.arcade.virtual.entity.display
 
 import com.mojang.math.Transformation
-import net.casual.arcade.networking.observer.tracker.ObserverTracker
-import net.casual.arcade.networking.utils.asPlayerOrNull
+import net.casual.arcade.observer.tracker.ObserverTracker
+import net.casual.arcade.observer.utils.asPlayerOrNull
 import net.casual.arcade.virtual.entity.SimpleVirtualEntity
 import net.casual.arcade.virtual.entity.attachment.VirtualEntityAttachment
 import net.minecraft.server.level.ServerPlayer
