+++ arcade-visuals/src/main/kotlin/net/casual/arcade/visuals/camera/PlayerCamera.kt
@@ -4,7 +4,7 @@
  */
 package net.casual.arcade.visuals.camera
 
-import net.casual.arcade.networking.utils.asObserver
+import net.casual.arcade.observer.utils.asObserver
 import net.casual.arcade.utils.ClientboundPlayerInfoUpdatePacket
 import net.casual.arcade.utils.asClientGamePacket
 import net.casual.arcade.utils.entity.teleportTo
