+++ arcade-visuals/src/main/kotlin/net/casual/arcade/visuals/entity/display/DynamicVirtualTextDisplay.kt
@@ -4,8 +4,8 @@
  */
 package net.casual.arcade.visuals.entity.display
 
-import net.casual.arcade.networking.observer.tracker.ObserverTracker
-import net.casual.arcade.networking.utils.asPlayerOrNull
+import net.casual.arcade.observer.tracker.ObserverTracker
+import net.casual.arcade.observer.utils.asPlayerOrNull
 import net.casual.arcade.virtual.entity.attachment.VirtualEntityAttachment
 import net.casual.arcade.virtual.entity.display.SimpleVirtualTextDisplay
 import net.casual.arcade.visuals.elements.PlayerSpecificElement
